feat: UIG-3053 - vl-functional-header - attribuut toegevoegd om sub-header te verbergen
Storybook verbeterd & cypress testen uitgebreid.
